pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
packages = [
    { include = "src" },
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "torch (>=2.7.0,<3.0.0)",
    "torchvision (>=0.22.0,<0.23.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "matplotlib (>=3.9,<4.0)",
    "diffusers (>=0.33.1,<0.34.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
